requirements.txt
anthropic>=0.40.0
polars>=1.0.0
duckdb>=1.0.0
geopandas>=1.0.0
shapely>=2.0.0
geopy>=2.4.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0
httpx>=0.27.0
Pillow>=10.0.0
python-dotenv>=1.0.0
streamlit>=1.38.0
pydeck>=0.9.0